Fix drones only loading one platform
When scanning platforms, if an in use platform had nothing to load the drone had already been assigned so we would not load the second platform. This moves the platform assignment under the part where we have things to taje. PLUS it moves the checkitems code to after we find no available platforms - this way we do not constantly rescan the items when drones are idle.