Updated controller code to use computronics (ep #43)
Updated code to use Computronics instead of transposers and engine detectors. 
Due to the change in colour numbers between computronics and the redstone this changes the previously saved state of engines so if upgrading from a previous version you are recommended to return all the engines back for reassignment, there is a warning generated when this is detected.